# proxy/internal/router/geo/docs/geo_swagger.yaml
basePath: /api
definitions:
  models.AddressGeo:
    description: AddressGeo represents the geocode result for an address.
    properties:
      suggestions:
        items:
          properties:
            data:
              properties:
                country:
                  type: string
                postal_code:
                  type: string
              type: object
            unrestricted_value:
              type: string
            value:
              type: string
          type: object
        type: array
    type: object
  models.AddressSearch:
    description: AddressSearch represents the search result for an address.
    properties:
      area: {}
      area_fias_id: {}
      area_kladr_id: {}
      area_type: {}
      area_type_full: {}
      area_with_type: {}
      beltway_distance: {}
      beltway_hit:
        type: string
      block: {}
      block_type: {}
      block_type_full: {}
      capital_marker:
        type: string
      city: {}
      city_area:
        type: string
      city_district:
        type: string
      city_district_fias_id: {}
      city_district_kladr_id: {}
      city_district_type:
        type: string
      city_district_type_full:
        type: string
      city_district_with_type:
        type: string
      city_fias_id: {}
      city_kladr_id: {}
      city_type: {}
      city_type_full: {}
      city_with_type: {}
      country:
        type: string
      country_iso_code:
        type: string
      entrance: {}
      federal_district:
        type: string
      fias_actuality_state:
        type: string
      fias_code:
        type: string
      fias_id:
        type: string
      fias_level:
        type: string
      flat:
        type: string
      flat_area:
        type: string
      flat_cadnum:
        type: string
      flat_fias_id:
        type: string
      flat_price:
        type: string
      flat_type:
        type: string
      flat_type_full:
        type: string
      floor: {}
      geo_lat:
        type: string
      geo_lon:
        type: string
      house:
        type: string
      house_cadnum:
        type: string
      house_fias_id:
        type: string
      house_kladr_id:
        type: string
      house_type:
        type: string
      house_type_full:
        type: string
      kladr_id:
        type: string
      metro:
        items:
          properties:
            distance:
              type: number
            line:
              type: string
            name:
              type: string
          type: object
        type: array
      okato:
        type: string
      oktmo:
        type: string
      postal_box: {}
      postal_code:
        type: string
      qc:
        type: integer
      qc_complete:
        type: integer
      qc_geo:
        type: integer
      qc_house:
        type: integer
      region:
        type: string
      region_fias_id:
        type: string
      region_iso_code:
        type: string
      region_kladr_id:
        type: string
      region_type:
        type: string
      region_type_full:
        type: string
      region_with_type:
        type: string
      result:
        type: string
      settlement: {}
      settlement_fias_id: {}
      settlement_kladr_id: {}
      settlement_type: {}
      settlement_type_full: {}
      settlement_with_type: {}
      source:
        type: string
      square_meter_price:
        type: string
      stead: {}
      stead_cadnum: {}
      stead_fias_id: {}
      stead_kladr_id: {}
      stead_type: {}
      stead_type_full: {}
      street:
        type: string
      street_fias_id:
        type: string
      street_kladr_id:
        type: string
      street_type:
        type: string
      street_type_full:
        type: string
      street_with_type:
        type: string
      tax_office:
        type: string
      tax_office_legal:
        type: string
      timezone:
        type: string
      unparsed_parts: {}
    type: object
  models.GeocodeRequest:
    description: GeocodeRequest represents the request body for address geocoding.
    properties:
      lat:
        type: string
      lon:
        type: string
    type: object
  models.SearchRequest:
    description: SearchRequest represents the request body for address search
    properties:
      query:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample API for address searching and geocoding using Dadata
    API.
  termsOfService: http://localhost:8080/geo/swagger/index.html
  title: GEO API
  version: "2.0"
paths:
  /address/geocode:
    post:
      consumes:
      - application/json
      operationId: GEO_address
      parameters:
      - description: Geocode request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GeocodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: get data
          schema:
            $ref: '#/definitions/models.AddressGeo'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Search for an GEO
      tags:
      - geo
  /address/search:
    post:
      consumes:
      - application/json
      operationId: search_address
      parameters:
      - description: Search request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: get data
          schema:
            $ref: '#/definitions/models.AddressSearch'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Search for an address
      tags:
      - geo
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"